requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
shap==0.42.1
lime==0.2.0.1
opencv-python==4.8.1.78
scikit-learn==1.3.1
scikit-image==0.21.0
pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2